fix(build): drop is_trivially_destructible static_assert and useless size cast
order.hpp still asserted std::is_trivially_destructible_v<order>; the previous commit only adjusted the comment. boost::intrusive::list_member_hook<normal_link> is trivially destructible on Apple libc++ but not on GNU libstdc++, so the assert was the actual cause of the Linux CMake matrix failures. Remove it; slab_arena calls ~order() on deallocate (a no-op for the hook in normal_link mode).

spsc_ring::size() had a static_cast<std::size_t> on an expression whose type already was std::size_t (uint64 minus uint64). GCC's -Wuseless-cast fires on the redundant cast. Drop it.

+++ include/lob/spsc_ring.hpp
@@ -63,7 +63,7 @@ class spsc_ring {
     [[nodiscard]] std::size_t size() const noexcept {
         const auto h = head_.load(std::memory_order_acquire);
         const auto t = tail_.load(std::memory_order_acquire);
-        return static_cast<std::size_t>(h - t);
+        return h - t;  // std::uint64_t -> std::size_t is a no-op on 64-bit targets
     }
 
     [[nodiscard]] bool empty() const noexcept { return size() == 0; }
